refactor: replase get request from App to Search component
