+++ src/App/App.tsx
@@ -5,14 +5,11 @@ import SearchResult from './components/SearchResult/SearchResult';
 import { Movie } from './components/SearchResult/types';
 import ShowMovieInfo from './components/SearchResult/ShowMovieInfo';
 import { connect } from 'react-redux'
-import { AppState, SearchQuery } from './typings/types';
-import { getMovies } from './thunkAction/getMovies';
-import { Dispatch, bindActionCreators } from 'redux';
+import { AppState} from './typings/types';
 import { ConnectedRouter } from 'connected-react-router';
 import { History } from 'history';
 import { Route, Switch } from 'react-router';
 import NotFound from './components/NotFound';
-import { history } from './ConfigureStore';
 
 type State = {
     showCurrentMovie: boolean;
@@ -21,8 +18,6 @@ type State = {
 
 type Props = {
     fetchedMovies: Movie[];
-    getMovies: Function;
-    searchQuery: SearchQuery;
     history: History;
 }
 
@@ -32,12 +27,6 @@ class App extends Component<Props, State> {
         this.state = {
             showCurrentMovie: false,
         }
-    }
-
-    componentDidMount() {
-        const url = `https://reactjs-cdp.herokuapp.com/movies?search=${history.location.pathname.substring(history.location.pathname.lastIndexOf('/') + 1).trim()}&searchBy=title`
-        console.log(url)
-        this.props.getMovies(url)
     }
 
     handleSelectMovie = (movie: Movie) => {
@@ -51,11 +40,6 @@ class App extends Component<Props, State> {
         this.setState({
             showCurrentMovie: false,
         });
-    }
-
-    handleSearch = () => {
-        const url = `https://reactjs-cdp.herokuapp.com/movies?search=${this.props.searchQuery.input}&searchBy=${this.props.searchQuery.filterOption}`
-        this.props.getMovies(url)
     }
     
     render() {
@@ -77,19 +61,13 @@ class App extends Component<Props, State> {
                             runtime = {this.state.currentMovie.runtime}
                         />
                         <SearchResult movies = {this.props.fetchedMovies} action = {(movie: Movie) => this.handleSelectMovie(movie)}/>
-                        <div className = 'footer'>
-                            <div className= 'footer-content'>Netflixroulette</div>
-                        </div>
                     </Route>
                 )
             }
             return (
                 <Route path="/" >
-                    <Search filterOptions = {['title', 'genres']} makeSearch = {this.handleSearch}/>
+                    <Search filterOptions = {['title', 'genres']} />
                     <SearchResult movies = {this.props.fetchedMovies} action = {(movie: Movie) => this.handleSelectMovie(movie)}/>
-                    <div className = 'footer'>
-                        <div className= 'footer-content'>Netflixroulette</div>
-                    </div>
                 </Route>
             );
         }
@@ -99,19 +77,17 @@ class App extends Component<Props, State> {
             <ConnectedRouter history={this.props.history}>
                 <Switch>
                     <Route exact path="/" >
-                        {/* <Route path="/search/:searchtext" component={Search} /> */}
-                        <Search filterOptions = {['title', 'genres']} makeSearch = {this.handleSearch}/>
-                        {/* <SearchResult movies = {this.props.fetchedMovies} action = {(movie: Movie) => this.handleSelectMovie(movie)}/> */}
+                        <Search filterOptions = {['title', 'genres']} />
                     </Route>
                     <Route path= "/search">
-                        <Search filterOptions = {['title', 'genres']} makeSearch = {this.handleSearch}/>
+                        <Search filterOptions = {['title', 'genres']} />
                         <SearchResult movies = {this.props.fetchedMovies} action = {(movie: Movie) => this.handleSelectMovie(movie)}/>
                     </Route>
                     <Route path="/films" >
                         {getResult()}
                     </Route>
                     <Route path="*" >
-                        <Search filterOptions = {['title', 'genres']} makeSearch = {this.handleSearch}/>
+                        <Search filterOptions = {['title', 'genres']} />
                         <NotFound />
                     </Route>
                 </Switch>
@@ -127,16 +103,9 @@ class App extends Component<Props, State> {
 function mapStateToProps (state: AppState){
     return {
         fetchedMovies: state.fetchedMovies,
-        searchQuery: state.searchQuery
-    }
-}
-function mapDispatchToProps (dispatch: Dispatch) {
-    return {
-        getMovies: bindActionCreators(getMovies, dispatch)
     }
 }
 
 export default connect(
-    mapStateToProps,
-    mapDispatchToProps
+    mapStateToProps
 )(App);
